Automatically enable relay when Internet is out
This completes the project. Every 5 minutes (or 1 minute when the internet
appears to be down), the device will poll to see if the internet is up. It
checks this by connecting to WiFi and sending a single ping to 8.8.8.8.

If 3 successive checks for internet connectivity fail, it is assumed that the
internet is down and the relay is activated (which will turn on the RG).